requirements.txt
Pillow==8.3.2
openai==0.27.0
requests
